Move all tests that regards the type fields parsing in the same module. And adjust the expected warnings line numbers (in some cases).
