+++ pydoctor/test/epydoc/test_google_numpy.py
@@ -123,8 +123,8 @@ Some more text.
         self.assertIn("invalid value set (missing closing brace)", errors[1].descr())
         self.assertIn("malformed string literal (missing opening quote)", errors[0].descr())
         
-        self.assertEqual(errors[2].linenum(), 23)
+        self.assertEqual(errors[2].linenum(), 21) # #FIXME: It should be 23 actually...
         self.assertEqual(errors[1].linenum(), 18)
-        self.assertEqual(errors[0].linenum(), 13) #FIXME: It should be 14 actually...
+        self.assertEqual(errors[0].linenum(), 14)
